Enlarge buffer of the file dumper by default
Change the size of the file dumper to 2MB. This saves us some context
switches for the read/write system calls and should improve performance
for large files.

We also introduce a configure switch (--disable-big-buffer) for memory
constrained systems. When this is set, we use the previous size of 16kB
instead.