Fix key_t signedness issues
key_t is signed on all supported systems, so fix our wrong assumption
for it being unsigned. We still support the unsigned octal and
hexadecimal syntaxes for the keys, so we need some traits of the key_t
data type. The code in keythelper.h does some guesses based on the size
of the key_t type.