shmcat.c, ftok.c: Terminate longopts correctly
The getopt_long(3) man page says we have to terminate the array of the
long options with a struct option containing all zeroes, so add the
missing termination. The problem was noticed because calling shmcat or
ftok with abbreviated long options (such as --ver instead of --version)
caused the program in question to segfault (see ticket #15). With this
commit abbreviated long options work as expected.